lib: main and run in bin, everything else in lib
